Create README.md
Curiosity gets the best of me as it always does.
Though as a bachelor of education, I need to be adaptable person and highly motivated to renew my brain's storage in order to be more knowledgeable. That's why I challenge my self to stay update so that Life can be more meaningful.
Meaningfulness is another name of being helpful and problem solver.

Stay hungry, and keep moving.